requirements.txt
bottle==0.12.13
certifi==2018.10.15
chardet==3.0.4
click==7.0
colorama==0.4.1
colorlog==4.0.2
configobj==5.0.6
Django==2.1.3
futures==3.1.1
idna==2.7
Jinja2==2.10
MarkupSafe==1.1.0
objectpath==0.6.1
pytz==2018.7
PyYAML==3.13
records==0.5.2
requests==2.20.1
selenium==3.141.0
tinydb==3.12.2
tomorrow==0.2.4
urllib3==1.24.1
